requirements.txt
# ARK - Answer-centric Retriever via KG-driven Curriculum Learning
# Python Dependencies

# Core Python version: >=3.8, <3.12

# ===== Deep Learning Frameworks =====
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
accelerate>=0.20.0
datasets>=2.14.0

# ===== Graph and Network Analysis =====
networkx>=3.0
graspologic>=3.2.0  # For stable largest connected component

# ===== Embedding and Vector DB =====
# Note: Ollama is used as external service (http://localhost:11434)
ollama-python>=0.1.0  # Optional Python client

# ===== LLM API Clients =====
openai>=1.0.0  # For OpenAI-compatible APIs
google-generativeai>=0.3.0  # For Gemini
anthropic>=0.7.0  # Optional: for Claude

# ===== Text Processing =====
tiktoken>=0.5.0  # OpenAI tokenizer
nltk>=3.8
spacy>=3.5.0

# ===== Scientific Computing =====
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0

# ===== Utilities =====
pyyaml>=6.0  # For config.yaml
python-dotenv>=1.0.0  # For environment variables
tqdm>=4.65.0
requests>=2.31.0
aiohttp>=3.8.0  # For async HTTP requests
tenacity>=8.2.0  # For retry logic

# ===== Token Cost Tracking =====
tokencost>=0.1.0  # Optional: for cost calculation

# ===== Logging and Monitoring =====
loguru>=0.7.0  # Enhanced logging
wandb>=0.15.0  # Optional: for experiment tracking

# ===== Evaluation Metrics =====
rouge-score>=0.1.2
sacrebleu>=2.3.0
bert-score>=0.3.13  # Optional: for BERTScore

# ===== Training Infrastructure =====
deepspeed>=0.12.0  # Optional: for distributed training
peft>=0.6.0  # Optional: for parameter-efficient fine-tuning
bitsandbytes>=0.41.0  # Optional: for quantization

# ===== MS-SWIFT (for Qwen training) =====
ms-swift>=2.0.0  # ModelScope Swift framework

# ===== Development and Testing =====
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.7.0  # Code formatting
flake8>=6.1.0  # Linting
mypy>=1.5.0  # Type checking
ipython>=8.14.0
jupyter>=1.0.0
notebook>=7.0.0

# ===== Data Processing =====
jsonlines>=4.0.0
msgpack>=1.0.5
orjson>=3.9.0  # Fast JSON parsing

# ===== Optional: Visualization =====
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# ===== Pin Critical Versions =====
setuptools>=65.0.0
wheel>=0.40.0
pip>=23.0.0

